fxi(coral): Add DetailsModalContent  component
